requirements.txt
streamlit
numpy
opencv-python
pillow
tensorflow
keras